docs(docsite): fix ColorVariants story to follow index re-export pattern
The 'Show code' panel was displaying the entire story module (meta,
argTypes and custom docs page) because ColorVariants/index.stories.tsx
defined the story inline instead of following the repo's index
re-export pattern.

Move the ColorIdPrefix story into a dedicated ColorVariants.stories.tsx
and re-export it from index.stories.tsx, matching the pattern used by
Icons/index.stories.tsx and FileTypeIcons/index.stories.tsx.
